files from Multi-Animal Tracker and saved outputs
these are the files iv'e been mostly messing with. The few changes i made to the original program are commented out for now (you can 'ctrl+f' search "NOTE" to find the places in the code where i commented). Although I'm pretty certain that at some point the program saves the shape/area of each track in regards to pixel-ID, it doesn't output it at the end and I haven't found a way to extract it yet. Next I'll probably try stepping through the program, line by line during runtime, to see if I can find it that way.